SCD2_process/write_to_staging/requirements.txt
Flask
google-cloud-storage
google-cloud-bigquery